requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
mysql-connector-python==8.2.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
Pillow==10.1.0
pydantic==2.5.0
passlib==1.7.4
bcrypt==4.1.1